+++ astropy/constants/si.py
@@ -114,6 +114,10 @@ kpc = Constant('kpc', "Kiloparsec",
                1000. * au.uncertainty / np.tan(np.radians(1. / 3600.)),
                "Derived from au", system='si')
 
+# Wien wavelength displacement law constant
+b_wien = Constant('b_wien', 'Wien wavelength displacement law constant',
+                  2.8977721e-3, 'm K', 0.0000026e-3, 'CODATA 2010', system='si')
+
 # SOLAR QUANTITIES
 
 # Solar luminosity